chore(acir)!: Remove `Block`, `RAM` and `ROM` opcodes (#457)
