+++ acvm/src/compiler/mod.rs
@@ -199,10 +199,6 @@ pub fn compile(
                 new_opcode_labels.push(opcode_label[index]);
                 transformed_gates.push(opcode.clone());
             }
-
-            Opcode::Block(_) | Opcode::ROM(_) | Opcode::RAM(_) => {
-                unimplemented!("Stepwise execution is not compatible with {}", opcode.name())
-            }
             Opcode::Brillig(brillig) => {
                 for output in &brillig.outputs {
                     match output {
